# perf: restore early stopping for _match [run_process_replay]