main: 
`n:  669639  tm: 480.27ms  tot: 718.49ms uopgraph.py:37:_match`
branch: 
`n:  669639  tm: 437.15ms  tot: 651.15ms uopgraph.py:37:_match`